+++ src/apps/seelen_wm/index.tsx
@@ -28,6 +28,7 @@ async function Main() {
       getCurrentWebviewWindow().show();
       let view = getCurrentWebviewWindow();
       view.emitTo(view.label, 'complete-setup');
+      view.emit('register-colors-events');
     }, []);
 
     return (