+++ src/apps/seelenweg/app.tsx
@@ -1,43 +1,36 @@
-import { ErrorBoundary } from '../seelenweg/components/Error';
 import { getRootContainer } from '../shared';
 import { useDarkMode } from '../shared/styles';
-import { ErrorFallback } from './components/Error';
+import { ErrorBoundary } from './components/Error';
+import { updateHitbox } from './events';
+import { SeelenWeg } from './modules/bar';
 import { emit, emitTo } from '@tauri-apps/api/event';
 import { getCurrentWebviewWindow } from '@tauri-apps/api/webviewWindow';
 import { ConfigProvider, theme } from 'antd';
 import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 
-import { ToolBar } from './modules/main/infra';
-
 import { Selectors } from './modules/shared/store/app';
 
 async function onMount() {
   let view = getCurrentWebviewWindow();
+  updateHitbox();
   await emitTo(view.label.replace('/', '-hitbox/'), 'init');
   await view.show();
+  await view.emitTo(view.label, 'complete-setup');
   await emit('register-colors-events');
 }
-
 export function App() {
-  const version = useSelector(Selectors.version);
-
-  const structure = useSelector(Selectors.placeholder);
+  const isDarkMode = useDarkMode();
   const colors = useSelector(Selectors.colors);
 
-  const isDarkMode = useDarkMode();
-
   useEffect(() => {
-    onMount().catch(console.error);
+    onMount();
   }, []);
-
-  if (!structure) {
-    return <ErrorFallback msg="NO PLACEHOLDER FOUND. PLEASE TRY REINSTALL THE APP" />;
-  }
 
   return (
     <ConfigProvider
-      getPopupContainer={() => getRootContainer()}
+      getPopupContainer={getRootContainer}
+      componentSize="small"
       theme={{
         token: {
           colorPrimary: isDarkMode ? colors.accent_light : colors.accent_dark,
@@ -45,8 +38,8 @@ export function App() {
         algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
       }}
     >
-      <ErrorBoundary fallback={<ErrorFallback />}>
-        <ToolBar key={version} structure={structure} />
+      <ErrorBoundary fallback={<div>Something went wrong</div>}>
+        <SeelenWeg />
       </ErrorBoundary>
     </ConfigProvider>
   );