+++ src/apps/seelenweg/modules/shared/store/app.ts
@@ -28,6 +28,18 @@ const initialState: RootState = {
   isOverlaped: false,
   settings: SeelenWegSlice.getInitialState(),
   mediaSessions: [],
+  colors: {
+    background: '#ffffff',
+    foreground: '#000000',
+    accent_darkest: '#000000',
+    accent_darker: '#000000',
+    accent_dark: '#000000',
+    accent: '#000000',
+    accent_light: '#000000',
+    accent_lighter: '#000000',
+    accent_lightest: '#000000',
+    complement: null,
+  },
 };
 
 function removeAppFromState(state: RootState, searched: SwPinnedApp | SwTemporalApp) {