+++ src/apps/seelenweg/modules/shared/store/domain.ts
@@ -57,6 +57,19 @@ export type SwTemporalApp = modify<
 
 export type SwItem = SwPinnedApp | SwTemporalApp | SavedSeparatorItem | SavedMediaItem | StartMenuItem;
 
+export interface UIColors {
+  background: string;
+  foreground: string;
+  accent_darkest: string;
+  accent_darker: string;
+  accent_dark: string;
+  accent: string;
+  accent_light: string;
+  accent_lighter: string;
+  accent_lightest: string;
+  complement: string | null;
+}
+
 export interface RootState extends IRootState<Seelenweg> {
   itemsOnLeft: SwItem[];
   itemsOnCenter: SwItem[];
@@ -67,4 +80,5 @@ export interface RootState extends IRootState<Seelenweg> {
   focusedExecutable: string;
   isOverlaped: boolean;
   mediaSessions: MediaSession[];
+  colors: UIColors;
 }