+++ src/apps/toolbar/modules/shared/store/app.ts
@@ -30,12 +30,23 @@ const initialState: RootState = {
   networkAdapters: [],
   networkLocalIp: null,
   online: false,
-  accentColor: '#ff0000',
   wlanBssEntries: [],
   mediaSessions: [],
   mediaOutputs: [],
   mediaInputs: [],
   notifications: [],
+  colors: {
+    background: '#ffffff',
+    foreground: '#000000',
+    accent_darkest: '#000000',
+    accent_darker: '#000000',
+    accent_dark: '#000000',
+    accent: '#000000',
+    accent_light: '#000000',
+    accent_lighter: '#000000',
+    accent_lightest: '#000000',
+    complement: null,
+  },
 };
 
 export const RootSlice = createSlice({