+++ src/background/modules/mod.rs
@@ -5,5 +5,6 @@ pub mod media;
 pub mod network;
 pub mod notifications;
 pub mod power;
+pub mod system_settings;
 pub mod tray;
 pub mod uwp;